refactor(topology): move classes related to making topology changes to own package
