+++ topology/src/main/java/io/camunda/zeebe/topology/changes/TopologyChangeAppliers.java
@@ -5,9 +5,10 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.topology;
+package io.camunda.zeebe.topology.changes;
 
 import io.camunda.zeebe.scheduler.future.ActorFuture;
+import io.camunda.zeebe.topology.state.ClusterTopology;
 import io.camunda.zeebe.topology.state.MemberState;
 import io.camunda.zeebe.topology.state.TopologyChangeOperation;
 import io.camunda.zeebe.util.Either;
@@ -29,7 +30,7 @@ public interface TopologyChangeAppliers {
      * operation. For example, an operation for joining a partition can mark the state as JOINING.
      *
      * @return an either which contains an exception if the operation is not valid, or a function to
-     *     update the cluster toplogy
+     *     update the cluster topology
      */
     Either<Exception, UnaryOperator<MemberState>> init();
